fixing an error with user vs sender
Summary:
just fixing an error for whether the object user or sender is called
when posting a message. i switched it to user bc sender was coming up null. what
is the point of sender anyway?? i think it's an unused object and we should
possibly delete it?

Test Plan: post a message -- the from field should no longer be null

Reviewers: ada, angelafz, wwwu

CC: trvish

Differential Revision: https://phabricator.fb.com/D912939